+++ pkg/iam/mocks.go
@@ -316,3 +316,49 @@ func (m *Mock) SetUserPassword(ctx context.Context, request SetUserPasswordReque
 
 	return args.Error(0)
 }
+
+func (m *Mock) ListProperties(ctx context.Context, request ListPropertiesRequest) (*ListPropertiesResponse, error) {
+	args := m.Called(ctx, request)
+
+	if args.Get(0) == nil {
+		return nil, args.Error(1)
+	}
+
+	return args.Get(0).(*ListPropertiesResponse), args.Error(1)
+}
+
+func (m *Mock) GetProperty(ctx context.Context, request GetPropertyRequest) (*GetPropertyResponse, error) {
+	args := m.Called(ctx, request)
+
+	if args.Get(0) == nil {
+		return nil, args.Error(1)
+	}
+
+	return args.Get(0).(*GetPropertyResponse), args.Error(1)
+}
+
+func (m *Mock) MoveProperty(ctx context.Context, request MovePropertyRequest) error {
+	args := m.Called(ctx, request)
+
+	return args.Error(0)
+}
+
+func (m *Mock) MapPropertyIDToName(ctx context.Context, request MapPropertyIDToNameRequest) (*string, error) {
+	args := m.Called(ctx, request)
+
+	if args.Get(0) == nil {
+		return nil, args.Error(1)
+	}
+
+	return args.Get(0).(*string), args.Error(1)
+}
+
+func (m *Mock) MapPropertyNameToID(ctx context.Context, request MapPropertyNameToIDRequest) (*int64, error) {
+	args := m.Called(ctx, request)
+
+	if args.Get(0) == nil {
+		return nil, args.Error(1)
+	}
+
+	return args.Get(0).(*int64), args.Error(1)
+}